discord_bot/requirements.txt
discord.py
python-dotenv
aiohttp
Pillow
beautifulsoup4
youtube-comment-downloader
google-genai
google-api-python-client
google-auth
asyncpg
psycopg2-binary
